% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj_heart.us.R
\name{rj_heart.us}
\alias{rj_heart.us}
\title{clear heart ultra sound data}
\usage{
rj_heart.us(dir)
}
\arguments{
\item{dir}{dir}
}
\value{
dataframe
}
\description{
clear heart ultra sound data
}
