% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj_labtest.getdata.R
\name{rj_labtest.getdata}
\alias{rj_labtest.getdata}
\title{get data for lab test}
\usage{
rj_labtest.getdata(dir)
}
\arguments{
\item{dir}{directory of labe test text}
}
\value{
df
}
\description{
get data for lab test
}
