% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj_labtest.query.R
\name{rj_labtest.query}
\alias{rj_labtest.query}
\title{query lab test data}
\usage{
rj_labtest.query(data, query)
}
\arguments{
\item{data}{data after rj_labtest.getdata}

\item{query}{one or more query words}
}
\value{
dataframe
}
\description{
query lab test data
}
